+++ example/files/main.dart
@@ -8,7 +8,6 @@ class MainFile extends Widget {
   Widget generate(Context context) {
     return For.of([
       // put your tick widgets here
-      Effect.clear(Entity(),EffectType.absorption)
     ]);
   }
 }