+++ lib/basic/widgets.dart
@@ -13,5 +13,6 @@ export 'entity.dart';
 export 'item.dart';
 export 'block.dart';
 export 'location.dart';
+export 'rotation.dart';
 export 'area.dart';
 export 'condition.dart';