+++ lib/utils/repeat.dart
@@ -1,0 +1,30 @@
+import 'package:meta/meta.dart';
+import 'package:objd/core.dart';
+
+ /// The Repeat Widget repeats a given action multiple times with a tick delay.
+class Repeat extends Widget {
+  String name;
+  String path;
+  Widget child;
+  String counter;
+  int to;
+  int ticks;
+  /// The Repeat Widget repeats a given action multiple times with a tick delay.
+  Repeat(this.name,{@required this.child,@required this.to,this.counter = "objd_repeat", this.path = "timers",this.ticks = 1});
+
+  @override
+  Widget generate(Context context) {
+      Score score = Score(Entity.PlayerName(name),counter);
+      String filePath = path + "/" + name;
+    return For.of([
+      score.set(0),
+      File.execute(path: filePath,child:For.of([
+        child,
+        score.add(1),
+        If(score.matchesRange(Range(to:to)),Then:[
+          Schedule(filePath,ticks: ticks)
+        ])
+      ]))
+    ]);
+  }
+}