+++ lib/utils/widgets.dart
@@ -1,2 +1,8 @@
 export 'timer.dart';
 export 'aroundlocation.dart';
+export 'log.dart';
+export 'armorstand.dart';
+export 'areaeffectcloud.dart';
+export 'hologram.dart';
+export 'randomScore.dart';
+export 'repeat.dart';