+++ lib/wrappers/effect.dart
@@ -1,6 +1,5 @@
 
 
-import 'package:meta/meta.dart';
 import 'package:objd/basic/command.dart';
 import 'package:objd/basic/entity.dart';
 import 'package:objd/basic/widget.dart';
@@ -25,7 +24,8 @@ class Effect extends Widget {
 /// )
 /// ⇒ effect give @p minecraft:jump_boost 20 3 true
 /// ```
-  Effect(this.effect,{@required this.entity,this.duration = 30,this.amplifier = 1,this.showParticles = true}){
+  Effect(this.effect,{this.entity,this.duration = 30,this.amplifier = 1,this.showParticles = true}){
+    if(entity == null) entity =Entity.Selected();
     _type = "give";
   }
 /// You can clear an effect again.
@@ -36,7 +36,14 @@ class Effect extends Widget {
   Effect.clear(this.entity,[this.effect]){
     _type = "clear";
   }
-
+  getMap(){
+    Map ret = {};
+    ret["Id"] = EffectType.values.indexOf(effect) + 1;
+    ret["Amplifier"] = amplifier != null ? amplifier - 1 : 0;
+    ret["Duration"] = duration != null ? duration * 20 : 0;
+    ret["ShowParticles"] = showParticles != null && showParticles ? 1 : 0;
+    return ret;
+  }
   @override
   Widget generate(Context context) {
     if(_type == "clear"){
@@ -48,33 +55,36 @@ class Effect extends Widget {
 }
 
 enum EffectType {
-  absorption,
-  unluck,
-  luck,
-  blindness,
-  conduit_power,
-  dolphins_grace,
-  fire_resistance,
-  glowing,
-  haste,
-  health_boost,
-  hunger,
-  instant_damage,
-  instant_health,
-  jump_boost,
-  levitation,
-  mining_fatigue,
-  nausea,
-  night_vision,
-  poison,
-  regeneration,
-  resistance,
-  saturation,
-  slow_falling,
-  slowness,
-  speed,
-  strength,
-  water_breathing,
-  weakness,
-  wither
+speed,
+slowness,
+haste,
+mining_fatique,
+strength,
+instant_health,
+instant_damage,
+jump_boost,
+nausea,
+regeneration,
+resistance,
+fire_resistance,
+water_breathing,
+invisibility,
+blindness,
+night_vision,
+hunger,
+weakness,
+poison,
+wither,
+health_boost,
+absorption,
+saturation,
+glowing,
+levitation,
+luck,
+unluck,
+slow_falling,
+conduit_power,
+dolphins_grace,
+bad_omen,
+
 }