+++ lib/wrappers/setblock.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:objd/basic/command.dart';
 import 'package:meta/meta.dart';
 import 'package:objd/basic/widget.dart';
@@ -8,11 +10,12 @@ import 'package:objd/build/build.dart';
 class SetBlock extends Widget {
   Location location;
   Block block;
+  Map<String,dynamic> nbt;
   /// The SetBlock Command Class sets a Block at the specified location.
-  SetBlock(this.block,{@required this.location});
+  SetBlock(this.block,{@required this.location,this.nbt});
   
   @override
   Widget generate(Context context){
-    return new Command('setblock ' + location.toString() + ' ' + block.toString());
+    return new Command('setblock ' + location.toString() + ' ' + block.toString() + (nbt == null? '': json.encode(nbt)));
   }
 }