+++ lib/wrappers/widgets.dart
@@ -17,3 +17,4 @@ export 'schedule.dart';
 export 'effect.dart';
 export 'particle.dart';
 export 'trigger.dart';
+export 'team.dart';